+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features (User Facing)
 * the print out of the Erlang version now is less verbose (just major/minor)
 * the fast_warning will now also tell you how to disable it
+* When `print: [benchmarking: false]` is set, information about which input is being benchmarked at the moment also won't be printed
 
 ## Breaking Changes (User Facing)
 * If you use the more verbose interface (`Benchee.init` and friends, e.g. not `Benchee.run`) then you have to insert a `Benchee.system` call before `Benchee.measure` (preferably right after `Benchee.init`)